# Reduced slow tasks in event_based_risk