--- a/debian/changelog
+++ b/debian/changelog
@@ -1,4 +1,5 @@
   [Michele Simionato]
+  * Strongly reduced the slow tasks in event_based_risk calculations
   * Fixed a performance issue in event based risk calculations, due to the same
     aggregation IDs being computed for each task
   * Increased `sys.recursionlimit` to 1500 to solve a rare pickling issue