--- a/openquake/calculators/event_based_risk.py
+++ b/openquake/calculators/event_based_risk.py
@@ -358,6 +358,7 @@ class EventBasedRiskCalculator(event_based.EventBasedCalculator):
                     self.datastore, oq.concurrent_tasks or 1)
                 allargs = [(arr, oq, self.datastore) for arr in slice_list]
             self.datastore.swmr_on()  # crucial!
+            logging.info('Starting ebr_from_gmfs')
             smap = parallel.Starmap(
                 ebr_from_gmfs, allargs, h5=self.datastore.hdf5)
             save_tmp(self, smap.monitor)