… openquake/commands/tests/commands_test.py