Supersedes https://github.com/gem/oq-engine/pull/8128. Part of https://github.com/gem/oq-engine/issues/8126. Here is the improvement for Argentina with 10,000 years on my workstation (over 2x speedup with half the memory used in "computing risk", over 2x in "reading data"):
```
# before
| calc_22361, maxmem=18.7 GB   | time_sec | memory_mb | counts |
|------------------------------+----------+-----------+--------|
| total event_based_risk       | 2_291    | 633.3     | 16     |
| computing risk               | 627.1    | 2_901     | 1_424  |
| EventBasedRiskCalculator.run | 462.9    | 1_837     | 1      |
| aggregating losses           | 128.0    | 0.0       | 1_936  |
| reading data                 | 49.3     | 736.4     | 16     |
| getting gmf_data slices      | 17.2     | 926.4     | 1      |

# after
| calc_22334, maxmem=16.6 GB   | time_sec | memory_mb | counts |
|------------------------------+----------+-----------+--------|
| total ebr_from_gmfs          | 998.5    | 675.8     | 17     |
| computing risk               | 590.7    | 1_624     | 1_513  |
| EventBasedRiskCalculator.run | 267.5    | 1_924     | 1      |
| aggregating losses           | 116.8    | 0.0       | 2_052  |
| reading data                 | 26.4     | 350.2     | 34     |
| getting gmf_data slices      | 17.6     | 927.1     | 1      |
```
The slow task situation is much better:
```
| operation-duration | counts | mean    | stddev | min     | max     | slowfac |
|--------------------+--------+---------+--------+---------+---------+---------|
| event_based_risk   | 16     | 143.2   | 25%    | 104.7   | 221.8   | 1.54922 |
| ebr_from_gmfs      | 17     | 58.7    | 26%    | 2.96108 | 73.7    | 1.25450 |
```
The approach scales really well even for large calculations, such as Chile with 100,000y on the spot machine:
```
# oq engine --run job_Chile.ini --hc 45118
| calc_45153, maxmem=281.6 GB  | time_sec | memory_mb | counts |
|------------------------------+----------+-----------+--------|
| total ebr_from_gmfs          | 83_102   | 736.6     | 1_010  |
| computing risk               | 66_592   | 1_921     | 71_710 |
| aggregating losses           | 7_088    | 0.0       | 88_329 |
| reading data                 | 3_354    | 718.9     | 2_020  |
| EventBasedRiskCalculator.run | 2_049    | 13_826    | 1      |
| averaging losses             | 962.7    | 0.0       | 88_329 |
| getting gmf_data slices      | 778.2    | 11_931    | 1      |

| operation-duration | counts | mean    | stddev | min     | max     | slowfac |
|--------------------+--------+---------+--------+---------+---------+---------|
| ebr_from_gmfs      | 1_010  | 82.3    | 15%    | 51.1    | 125.5   | 1.52563 |
```
**We are using 10 times less memory than before in "computing risk"!**